Upgrade to langchain v0.0.344 and langchain-experimental v0.0.43

+++ tests/__init__.py
@@ -17,4 +17,5 @@ class TemporaryCassettePath:
             assert os.path.isfile(self.cassette_path)
         finally:
             # remove it for future testing
-            os.remove(self.cassette_path)
+            if os.path.isfile(self.cassette_path):
+                os.remove(self.cassette_path)
